// src/app/models/condition/LogicalOperator.ts
export enum LogicalOperator {
  // Logical
  AND = 'AND',
  OR = 'OR',
  XOR = 'XOR',

  // Comparison
  EQUALS = 'EQUALS',
  NOT_EQUALS = 'NOT_EQUALS',
  GREATER_THAN = 'GREATER_THAN',
  LESS_THAN = 'LESS_THAN',
  GREATER_THAN_OR_EQUAL = 'GREATER_THAN_OR_EQUAL',
  LESS_THAN_OR_EQUAL = 'LESS_THAN_OR_EQUAL'
}